Fix 'Not the right place?' tap target on trip page
The trip page (TripDetailView) wrapped each visit row in a NavigationLink
inside a ScrollView, which swallowed taps on the embedded 'Not the right
place?' button. It also never passed the onPickAlternative closure nor
presented the alternative-place sheet, so the affordance was inert there.

- Extract AlternativePlacePicker into its own file so it can be reused
  outside LogbookView.
- In TripDetailView, navigate to PlaceDetailView via a plain Button +
  navigationDestination instead of a row-wrapping NavigationLink, which
  lets the inner button receive its own taps.
- Wire up onPickAlternative and present AlternativePlacePicker as a sheet.
